+++ src/serializer/log/metablock_manager.tcc
@@ -14,7 +14,7 @@
 
 template<class metablock_t>
 metablock_manager_t<metablock_t>::metablock_manager_t::head_t::head_t(metablock_manager_t *manager)
-    : mb_slot(0), saved_mb_slot(-1), wraparound(false), mgr(manager) { }
+    : mb_slot(0), saved_mb_slot(static_cast<uint32_t>(-1)), wraparound(false), mgr(manager) { }
 
 template<class metablock_t>
 void metablock_manager_t<metablock_t>::metablock_manager_t::head_t::operator++(UNUSED int stupid_plusplus_parameter) {
@@ -40,9 +40,9 @@ void metablock_manager_t<metablock_t>::metablock_manager_t::head_t::push() {
 
 template<class metablock_t>
 void metablock_manager_t<metablock_t>::metablock_manager_t::head_t::pop() {
-    guarantee(saved_mb_slot != (uint32_t) -1, "Popping without a saved state");
+    guarantee(saved_mb_slot != static_cast<uint32_t>(-1), "Popping without a saved state");
     mb_slot = saved_mb_slot;
-    saved_mb_slot = -1;
+    saved_mb_slot = static_cast<uint32_t>(-1);
 }
 
 template<class metablock_t>
